daml-script/runner: extract RunnerTestBase from RunnerMainTestBase
